don't hold lock until printing
I forgot that the out << rule situation does all the calculation
at printing time. So we're grabbing the lock way too early. Instead,
we will dump everything in a stringstream and then only grab the
lock when we need to go to stdout.

There was a linear speedup on a small test on my laptop.